What did I just make
The important thing isn't the oscillating squares.
Still, it looks siiiiiiiick. I need to see if this runs full speed on
android.

However, I made a thingy where the location of a drawing is based on the
PerspectiveCamera, while its rendered using the OrthographicCamera. Then
I tried to make a dance party, and then I got this.