Decent product so far.
Two issues still remain. FOV messes up the units, so what would be a
world unit would be something more or less. Secondly, the sprites
orientation doesn't reflect the z axis at that point. I can get the
angle, but cannot get the "feet" of the sprite to match the correct
position. :/